+++ apps/vscode/webview-ui/src/components/chat/chat-view/hooks/index.ts
@@ -3,5 +3,6 @@
  */
 
 export { useChatState } from "./useChatState"
+export { isThinkingPlaceholderRow, THINKING_PLACEHOLDER_TS, useDisplayedGroupedMessages } from "./useDisplayedMessages"
 export { useMessageHandlers } from "./useMessageHandlers"
 export { useScrollBehavior } from "./useScrollBehavior"
